Renamed LeftSidebar to EditInformation
