+++ src/modules/customers/components/detail/sidebar/EditInformation.js
