+++ src/modules/customers/components/detail/sidebar/index.js
@@ -3,7 +3,7 @@ import MessengerSection from './MessengerSection';
 import TwitterSection from './TwitterSection';
 import FacebookSection from './FacebookSection';
 import BasicInfo from './BasicInfo';
-import LeftSidebar from './LeftSidebar';
+import EditInformation from './EditInformation';
 
 export {
   FacebookSection,
@@ -11,5 +11,5 @@ export {
   MessengerSection,
   TaggerSection,
   BasicInfo,
-  LeftSidebar
+  EditInformation
 };
